refactor(subagents-v3): rename subagents files
