+++ packages/pi-subagents-v3/src/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./subagents-v3.js";
+export { default } from "./subagents.js";
