+++ packages/pi-subagents-v3/src/subagents.ts
